General: Reuse the edge auto-scroll for other drag gestures
The auto-scroller that keeps a drag-select range growing while the finger
rests at a viewport edge is not specific to selection: it needs a scrollable
container, an edge distance and a per-frame callback. Moved it out of
DragSelect.kt into common/compose/autoscroll as EdgeAutoScroller plus the
AutoScrollTarget abstraction over LazyList/LazyGrid, with the edge distance
and top speed as companion defaults. Drag-select behaviour is unchanged.
